+++ sources/Scripts/AsteroidController.hpp
@@ -3,7 +3,7 @@
 class AsteroidController : public ScriptBase
 {
   public:
-    explicit AsteroidController(BaseScene* scene, unsigned owner_ID) : ScriptBase(scene, owner_ID)
+    explicit AsteroidController(BaseScene* scene, unsigned short owner_ID) : ScriptBase(scene, owner_ID)
     {
 	printf("AsteroidController created\n");
     }
