+++ sources/Scripts/EnemyShipController.hpp
@@ -3,7 +3,7 @@
 class EnemyShipController : public ScriptBase
 {
   public:
-    explicit EnemyShipController(BaseScene* scene, unsigned owner_ID) : ScriptBase(scene, owner_ID)
+    explicit EnemyShipController(BaseScene* scene, unsigned short owner_ID) : ScriptBase(scene, owner_ID)
     {
     }
   private:
