+++ sources/Scripts/HealthController.hpp
@@ -3,7 +3,7 @@
 class HealthController : public ScriptBase
 {
   public:
-	explicit HealthController(BaseScene* scene, unsigned owner_ID) : ScriptBase(scene, owner_ID)
+	explicit HealthController(BaseScene* scene, unsigned short owner_ID) : ScriptBase(scene, owner_ID)
 	{
 	}
   private:
